+++ src/intel.cxx
@@ -1,0 +1,5 @@
+#include "intel.h"
+
+void IntelFortranInvocation::loadToolchainDependentSpackVars(SpackEnvState &spackenv) {
+    this->spackCommand = spackenv.SpackFC;
+}
