+++ src/ld.cxx
@@ -1,0 +1,5 @@
+#include "ld.h"
+
+void LdInvocation::loadToolchainDependentSpackVars(SpackEnvState &spackenv) {
+    this->spackCommand = std::string("link.exe");
+}
