+++ src/utils.h
@@ -6,6 +6,40 @@
 #include <strsafe.h>
 
 typedef std::vector<std::string> StrList;
+
+class SpackException : public std::exception {
+public:
+    SpackException(std::string msg) : message(msg) {}
+    char const* what();
+protected:
+    std::string message;
+};
+
+class SpackUnknownCompilerException : public SpackException {
+    using SpackException::SpackException;
+public:
+    char const * what();
+};
+
+class SpackCompilerException : public SpackException {
+    using SpackException::SpackException;
+public:
+    char const * what();
+};
+
+class SpackCompilerContextException : public SpackException {
+    using SpackException::SpackException;
+public:
+    char const * what();
+};
+
+
+// Environment Helper Methods
+std::string getSpackEnv(const char* env);
+std::string getSpackEnv(std::string env);
+StrList getenvlist(std::string envVar, std::string delim = ";");
+void validate_spack_env();
+
 // String helper methods adding cxx20 features to cxx14
 
 // Returns true if arg starts with match
@@ -32,6 +66,8 @@ std::wstring ConvertAnsiToWide(const std::string &str);
 StrList split(std::string s, std::string delim);
 
 // Parse command line opts
-bool isPatch(const char ** begin, const char ** end);
+bool isPatch(const char * arg);
 
-StrList parsePatch(const char ** args);
+std::map<std::string, std::string> parsePatch(const char ** args, int argc);
+
+bool checkAndPrintHelp(const char * arg);
